Add files via upload
 Write a C program to do the following:
a.      consider the customer name, customer number, customer type and number of overdue days (for return the book) in a library. Read customer number as a 10 digit number, Customer type as S or F to indicate Student or Faculty and number of days late after the due date. Display details in a formatted way.
b.    Calculate the fine amount as follows
For customer type as S, the rate is given as
For first 5 days Rs. 2per day
For next 5 days Rs.2.5 per day
For next 10 days Rs. 3.25 per day
An additional charge of Rs.1000 is charged if book is damaged.
For customer type as F, the rate is given as
For first 5 days Rs. 4.5per day
For next 5 days Rs.7.5 per day
For next 10 days Rs. 10.5 per day
An additional charge of Rs.2000 is charged if book is damaged.
Display bill in a formatted way.